refactor(core): move process manager into root
