+++ packages/core/src/shared/stop.ts
@@ -1,6 +1,6 @@
 import cli from "cli-ux";
 import { BaseCommand } from "../commands/command";
-import { processManager } from "../services/process-manager";
+import { processManager } from "../process-manager";
 
 export abstract class AbstractStopCommand extends BaseCommand {
     public async run(): Promise<void> {
